Failed attempts at sharing things
